Lift the runner logic too
